+++ examples/todo/tests/src/todo.spec.ts
@@ -10,11 +10,10 @@ const WASM_PATH = resolve(
   '..',
   '..',
   '..',
-  '.dfx',
-  'local',
-  'canisters',
+  '.icp',
+  'cache',
+  'artifacts',
   'todo',
-  'todo.wasm.gz',
 );
 
 describe('Todo', () => {